server/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
